pyproject.toml
[tool.poetry]
name = "django-compressor-js"
version = "0.1.0"
description = "Transpile JavaScript without running an extra node.js process."
authors = ["Adam Hill <adamghill@yahoo.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/adamghill/django-compressor-js"
keywords = ["django", "django-compressor", "babel", "javascript"]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.6"
dukpy = "^0.2.2"
django-compressor = "^2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
